Chore: Pulling Cpp to Go adapter into its own file
